Extending fonts make for Reverie rendering engine integration on Android 4.X
Reverie rendering engine offers high accuracy rendering and fonts for
various world languages. This project is a part of the integration of
Reverie rendering engine with Android version 4.1.2 and 4.2.2.

These modifications can be enabled through pre-processor directive flag
"REVERIE". The flag can be defined in the make files for compiling the
respective source files. If REVERIE is not defined, the files can be
compiled as and with the standard original project source for the
respective versions. If REVERIE is defined, the project compilation will
require additional reverie proprietary modules at link time for a
successfully integrated build with Reverie solution. These additional
proprietary modules are properties of Reverie and are not open sourced.